sunxi-mali: Fix build of test for case DISTRO_FEATURES_remove = "x11 wayland".
As eglplatform_fb.h is moved into image/sysroot as eglplatform.h, we need to
build the test application against this renamed include header, which resides
in the package image folder. Thus modify the include path in the patch.

This also removes the hard-coded linking against -lX11. It is not yet tested
if this is linking is covered by the Makefiles already, or need to be fixed
up depending on PACKAGECONF in the meta-sunxi recipe.

(Another option would be to build the test application only after the libraries
have been staged into the target sysroot.)

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>